requirements-raspberry-pi.txt
# Human Tracking Car Requirements - Raspberry Pi OS
# Use this file for installation on Raspberry Pi OS

# Core web framework
flask>=2.0.0

# Mathematical operations
numpy>=1.21.0

# Image processing utilities
imutils>=0.5.4

# NOTE: For OpenCV on Raspberry Pi, it's recommended to install via system packages:
# sudo apt install python3-opencv
# opencv-python  # Uncomment only if system package doesn't work

# GPIO control for Raspberry Pi
RPi.GPIO>=0.7.0
gpiozero>=1.6.2

# NOTE: For camera on Raspberry Pi OS, install via system packages:
# sudo apt install python3-picamera2
# picamera  # Legacy camera library (uncomment if needed)

# I2C and PWM control
adafruit-circuitpython-pca9685>=3.4.0
adafruit-circuitpython-motor>=3.4.0
adafruit-blinka>=8.0.0

# Optional: pygame for additional input handling
pygame>=2.1.0

# System utilities
psutil>=5.8.0